enable previously failing tests
