--- src/test/java/com/fasterxml/jackson/failing/FailingNonStandardNumberParsing782Test.java
@@ -1,44 +1,0 @@
-package com.fasterxml.jackson.failing;
-
-import com.fasterxml.jackson.core.BaseTest;
-import com.fasterxml.jackson.core.JsonFactory;
-import com.fasterxml.jackson.core.JsonParser;
-import com.fasterxml.jackson.core.JsonToken;
-import com.fasterxml.jackson.core.json.JsonReadFeature;
-
-public class FailingNonStandardNumberParsing782Test
-    extends BaseTest
-{
-    private final JsonFactory JSON_F = JsonFactory.builder()
-            .enable(JsonReadFeature.ALLOW_LEADING_PLUS_SIGN_FOR_NUMBERS)
-            .enable(JsonReadFeature.ALLOW_LEADING_DECIMAL_POINT_FOR_NUMBERS)
-            .enable(JsonReadFeature.ALLOW_TRAILING_DECIMAL_POINT_FOR_NUMBERS)
-            .build();
-
-    protected JsonFactory jsonFactory() {
-        return JSON_F;
-    }
-
-    public void testLeadingDotInNegativeDecimalAllowedAsync() throws Exception {
-        _testLeadingDotInNegativeDecimalAllowed(jsonFactory(), MODE_DATA_INPUT);
-    }
-
-    public void testLeadingDotInNegativeDecimalAllowedBytes() throws Exception {
-        _testLeadingDotInNegativeDecimalAllowed(jsonFactory(), MODE_INPUT_STREAM);
-        _testLeadingDotInNegativeDecimalAllowed(jsonFactory(), MODE_INPUT_STREAM_THROTTLED);
-    }
-
-    public void testLeadingDotInNegativeDecimalAllowedReader() throws Exception {
-        _testLeadingDotInNegativeDecimalAllowed(jsonFactory(), MODE_READER);
-    }
-
-    private void _testLeadingDotInNegativeDecimalAllowed(JsonFactory f, int mode) throws Exception
-    {
-        try (JsonParser p = createParser(f, mode, " -.125 ")) {
-            assertEquals(JsonToken.VALUE_NUMBER_FLOAT, p.nextToken());
-            assertEquals(-0.125, p.getValueAsDouble());
-            assertEquals("-0.125", p.getDecimalValue().toString());
-            assertEquals("-.125", p.getText());
-        }
-    }
-}
